+++ src/lib/ReactDataTable/ReactDataTable.tsx
@@ -22,6 +22,7 @@ const ReactDataTable = <TData,>(props: ReactDataTableProps<TData>) => {
     rowStyle,
     pageSizes,
     showPaging,
+    hidePageSizeChange,
     onEnter,
     totalRecords = table.getCoreRowModel().rows.length,
     withoutHeaders = false,
@@ -218,6 +219,7 @@ const ReactDataTable = <TData,>(props: ReactDataTableProps<TData>) => {
             showedItemsText: reactDataTableTranslations.showedItemsText,
           }}
           pagingPossible={true}
+          changePageSizePossible={!hidePageSizeChange}
         />
       )}
     </>