+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- option `hidePageSizeChange` added to hide the possibility to change the page size
+
 ## [4.0.0] - 2024-03-06
 
 ### Added